requirements-dev.txt
# requirements-dev.txt
ansible-lint
ruamel.yaml
black
pre-commit
yamllint
poetry
